+++ cadwyn/codegen/_asts.py
@@ -2,6 +2,7 @@ import ast
 import inspect
 import re
 from collections.abc import Callable
+from dataclasses import dataclass
 from enum import Enum, auto
 from pathlib import Path
 from types import GenericAlias, LambdaType, ModuleType, NoneType
@@ -15,8 +16,7 @@ from typing import (  # noqa: UP035
 
 from cadwyn._compat import (
     PYDANTIC_V2,
-    get_attrs_that_are_not_from_field_and_that_are_from_field,
-    is_pydantic_constrained_type,
+    is_pydantic_1_constrained_type,
 )
 from cadwyn._package_utils import (
     get_absolute_python_path_of_import,
@@ -108,29 +108,22 @@ def transform_none(_: NoneType) -> Any:
 
 def transform_type(value: type) -> Any:
     # This is a hack for pydantic's Constrained types
-    if is_pydantic_constrained_type(value):
-        if get_attrs_that_are_not_from_field_and_that_are_from_field(value)[0]:
-            parent = value.mro()[1]
-            snake_case = _RE_CAMEL_TO_SNAKE.sub("_", value.__name__)
-            cls_name = "con" + "".join(snake_case.split("_")[1:-1])
-            return (
-                cls_name.lower()
-                + "("
-                + ", ".join(
-                    [
-                        f"{key}={get_fancy_repr(val)}"
-                        for key, val in value.__dict__.items()
-                        if not key.startswith("_") and val is not None and val != parent.__dict__[key]
-                    ],
-                )
-                + ")"
+    if is_pydantic_1_constrained_type(value):
+        parent = value.mro()[1]
+        snake_case = _RE_CAMEL_TO_SNAKE.sub("_", value.__name__)
+        cls_name = "con" + "".join(snake_case.split("_")[1:-1])
+        return (
+            cls_name.lower()
+            + "("
+            + ", ".join(
+                [
+                    f"{key}={get_fancy_repr(val)}"
+                    for key, val in value.__dict__.items()
+                    if not key.startswith("_") and val is not None and val != parent.__dict__[key]
+                ],
             )
-        else:
-            # In pydantic V1:
-            # MRO of constr looks like: [ConstrainedStrValue, pydantic.types.ConstrainedStr, str, object]
-            #                                                                                -2     -1
-            #                                                                                ^^^
-            value = value.mro()[-2]
+            + ")"
+        )
 
     return value.__name__
 
@@ -230,11 +223,17 @@ def add_keyword_to_call(attr_name: str, attr_value: Any, call: ast.Call):
         call.keywords.append(new_keyword)
 
 
+def delete_keyword_from_call(attr_name: str, call: ast.Call):
+    for i, keyword in enumerate(call.keywords):  # pragma: no branch
+        if keyword.arg == attr_name:
+            call.keywords.pop(i)
+            break
+
+
 def get_ast_keyword_from_argument_name_and_value(name: str, value: Any):
-    return ast.keyword(
-        arg=name,
-        value=ast.parse(get_fancy_repr(value), mode="eval").body,
-    )
+    if not isinstance(value, ast.AST):
+        value = ast.parse(get_fancy_repr(value), mode="eval").body
+    return ast.keyword(arg=name, value=value)
 
 
 def pop_docstring_from_cls_body(cls_body: list[ast.stmt]) -> list[ast.stmt]:
@@ -247,3 +246,31 @@ def pop_docstring_from_cls_body(cls_body: list[ast.stmt]) -> list[ast.stmt]:
         return [cls_body.pop(0)]
     else:
         return []
+
+
+@dataclass(slots=True)
+class _ValidatorWrapper:
+    func_ast: ast.FunctionDef
+    index_of_validator_decorator: int
+    field_names: set[str | ast.expr] | None
+    is_deleted: bool = False
+
+
+def get_validator_info_or_none(method: ast.FunctionDef) -> _ValidatorWrapper | None:
+    for index, decorator in enumerate(method.decorator_list):
+        # The cases we handle here:
+        # * `Name(id="root_validator")`
+        # * `Call(func=Name(id="validator"), args=[Constant(value="foo")])`
+        # * `Attribute(value=Name(id="pydantic"), attr="root_validator")`
+        # * `Call(func=Attribute(value=Name(id="pydantic"), attr="root_validator"), args=[])`
+
+        if isinstance(decorator, ast.Call) and ast.unparse(decorator.func).endswith("validator"):
+            if len(decorator.args) == 0:
+                return _ValidatorWrapper(method, index, None)
+            else:
+                return _ValidatorWrapper(
+                    method, index, {arg.value if isinstance(arg, ast.Constant) else arg for arg in decorator.args}
+                )
+        elif isinstance(decorator, ast.Name | ast.Attribute) and ast.unparse(decorator).endswith("validator"):
+            return _ValidatorWrapper(method, index, None)
+    return None