+++ cadwyn/codegen/_common.py
@@ -16,6 +16,8 @@ from cadwyn._package_utils import IdentifierPythonPath
 from cadwyn.exceptions import CodeGenerationError
 from cadwyn.structure.versions import Version
 
+from ._asts import _ValidatorWrapper, get_validator_info_or_none
+
 _FieldName: TypeAlias = str
 _CodegenPluginASTType = TypeVar("_CodegenPluginASTType", bound=ast.AST)
 
@@ -25,6 +27,7 @@ class PydanticModelWrapper:
     cls: type[BaseModel]
     name: str
     fields: dict[_FieldName, PydanticFieldWrapper]
+    validators: dict[_FieldName, _ValidatorWrapper]
     _parents: list[Self] | None = dataclasses.field(init=False, default=None)
 
     def _get_parents(self, schemas: "dict[IdentifierPythonPath, Self]"):
@@ -37,11 +40,14 @@ class PydanticModelWrapper:
             if schema_path in schemas:
                 parents.append(schemas[schema_path])
             elif issubclass(base, BaseModel):
-                parents.append(type(self)(base, base.__name__, get_fields_from_model(base)))
+                fields, validators = get_fields_and_validators_from_model(base)
+                parents.append(type(self)(base, base.__name__, fields, validators))
         self._parents = parents
         return parents
 
-    def _get_defined_fields(self, schemas: "dict[IdentifierPythonPath, Self]") -> dict[str, PydanticFieldWrapper]:
+    def _get_defined_fields_through_mro(
+        self, schemas: "dict[IdentifierPythonPath, Self]"
+    ) -> dict[str, PydanticFieldWrapper]:
         fields = {}
 
         for parent in reversed(self._get_parents(schemas)):
@@ -49,9 +55,19 @@ class PydanticModelWrapper:
 
         return fields | self.fields
 
+    def _get_defined_annotations_through_mro(self, schemas: "dict[IdentifierPythonPath, Self]") -> dict[str, Any]:
+        annotations = {}
+
+        for parent in reversed(self._get_parents(schemas)):
+            annotations |= parent.cls.__annotations__
+
+        return annotations | self.cls.__annotations__
+
 
 @cache
-def get_fields_from_model(cls: type) -> dict[str, PydanticFieldWrapper]:
+def get_fields_and_validators_from_model(
+    cls: type,
+) -> tuple[dict[_FieldName, PydanticFieldWrapper], dict[_FieldName, _ValidatorWrapper]]:
     if not isinstance(cls, type) or not issubclass(cls, BaseModel):
         raise CodeGenerationError(f"Model {cls} is not a subclass of BaseModel")
 
@@ -59,25 +75,37 @@ def get_fields_from_model(cls: type) -> dict[str, PydanticFieldWrapper]:
     try:
         source = inspect.getsource(cls)
     except OSError:
-        return {
-            field_name: PydanticFieldWrapper(
-                annotation=field.annotation,
-                init_model_field=field,
-            )
-            for field_name, field in fields.items()
-        }
+        return (
+            {
+                field_name: PydanticFieldWrapper(annotation=field.annotation, init_model_field=field)
+                for field_name, field in fields.items()
+            },
+            {},
+        )
     else:
         cls_ast = cast(ast.ClassDef, ast.parse(source).body[0])
-        return {
-            node.target.id: PydanticFieldWrapper(
-                annotation=fields[node.target.id].annotation,
-                init_model_field=fields[node.target.id],
-                annotation_ast=node.annotation,
-                field_ast=node.value,
-            )
+        validators: dict[str, _ValidatorWrapper] = {}
+
+        validators_and_nones = (
+            get_validator_info_or_none(node)
             for node in cls_ast.body
-            if isinstance(node, ast.AnnAssign) and isinstance(node.target, ast.Name) and node.target.id in fields
-        }
+            if isinstance(node, ast.FunctionDef) and node.decorator_list
+        )
+        validators = {validator.func_ast.name: validator for validator in validators_and_nones if validator is not None}
+
+        return (
+            {
+                node.target.id: PydanticFieldWrapper(
+                    annotation=fields[node.target.id].annotation,
+                    init_model_field=fields[node.target.id],
+                    annotation_ast=node.annotation,
+                    value_ast=node.value,
+                )
+                for node in cls_ast.body
+                if isinstance(node, ast.AnnAssign) and isinstance(node.target, ast.Name) and node.target.id in fields
+            },
+            validators,
+        )
 
 
 @dataclass(slots=True)