+++ tests/test_routing.py
@@ -9,6 +9,7 @@ from typing import Annotated, Any, NewType, TypeAlias, cast, get_args
 from uuid import UUID
 
 import pytest
+from dirty_equals import IsStr
 from fastapi import APIRouter, Body, Depends, UploadFile
 from fastapi.routing import APIRoute
 from fastapi.testclient import TestClient
@@ -959,7 +960,7 @@ def test__router_generation__updating_request_depends(
                     "loc": ["body", "foo"],
                     "msg": "Field required",
                     "input": {},
-                    "url": "https://errors.pydantic.dev/2.5/v/missing",
+                    "url": IsStr,
                 },
             ],
         }
@@ -970,7 +971,7 @@ def test__router_generation__updating_request_depends(
                     "loc": ["body", "foo"],
                     "msg": "Field required",
                     "input": {},
-                    "url": "https://errors.pydantic.dev/2.5/v/missing",
+                    "url": IsStr,
                 },
             ],
         }